remove old region code
